# strava-backend/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ApiResponse_any_": {
				"properties": {
					"data": {
						"items": {},
						"type": "array"
					},
					"status": {
						"properties": {
							"message": {
								"type": "string"
							},
							"success": {
								"type": "boolean"
							}
						},
						"required": [
							"message",
							"success"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PAGINATION_ORDER": {
				"enum": [
					"ASC",
					"DESC"
				],
				"type": "string"
			},
			"DataTableQuery": {
				"properties": {
					"q": {
						"type": "string"
					},
					"field": {
						"type": "string"
					},
					"order": {
						"$ref": "#/components/schemas/PAGINATION_ORDER"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"perPage": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Achievement": {
				"properties": {
					"max_users": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"imagePath": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"max_users",
					"description",
					"imagePath",
					"title",
					"id"
				],
				"type": "object",
				"description": "Model Achievement"
			},
			"ApiResponse_Achievement-Array_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Achievement"
						},
						"type": "array"
					},
					"status": {
						"properties": {
							"message": {
								"type": "string"
							},
							"success": {
								"type": "boolean"
							}
						},
						"required": [
							"message",
							"success"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FilteredUserInterface": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthenticationServiceTokenResponseData": {
				"properties": {
					"token": {
						"type": "string"
					},
					"expires_in": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/FilteredUserInterface"
					}
				},
				"required": [
					"token",
					"expires_in",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_AuthenticationServiceTokenResponseData_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/AuthenticationServiceTokenResponseData"
					},
					"status": {
						"properties": {
							"message": {
								"type": "string"
							},
							"success": {
								"type": "boolean"
							}
						},
						"required": [
							"message",
							"success"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserLoginRequest": {
				"properties": {
					"email": {
						"type": "string",
						"format": "email"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"roleId": {
						"type": "number",
						"format": "double"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"roleId",
					"password",
					"name",
					"email",
					"id"
				],
				"type": "object",
				"description": "Model User"
			},
			"ApiResponse_User_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/User"
					},
					"status": {
						"properties": {
							"message": {
								"type": "string"
							},
							"success": {
								"type": "boolean"
							}
						},
						"required": [
							"message",
							"success"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthScope": {
				"enum": [
					"Admin",
					"User"
				],
				"type": "string"
			},
			"UserCreateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"scope": {
						"$ref": "#/components/schemas/AuthScope"
					}
				},
				"required": [
					"name",
					"email",
					"password",
					"scope"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserViewModel": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string",
						"format": "email"
					}
				},
				"required": [
					"id",
					"name",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserViewModel_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/UserViewModel"
					},
					"status": {
						"properties": {
							"message": {
								"type": "string"
							},
							"success": {
								"type": "boolean"
							}
						},
						"required": [
							"message",
							"success"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "rest-express",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/v1/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "field",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PAGINATION_ORDER"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "perPage",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/achievements": {
			"get": {
				"operationId": "GetAllAchievements",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Achievement-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Achievement"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "field",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PAGINATION_ORDER"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "perPage",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/auth/login/password": {
			"post": {
				"operationId": "LoginPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_AuthenticationServiceTokenResponseData_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserLoginRequest"
							}
						}
					}
				}
			}
		},
		"/auth/tokens/refresh": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_AuthenticationServiceTokenResponseData_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": []
			}
		},
		"/auth/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_User_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreateRequest"
							}
						}
					}
				}
			}
		},
		"/auth/user": {
			"get": {
				"operationId": "User",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserViewModel_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}